refactor: extract Redfish code from carbide-api into carbide-redfish
Part of an ongoing effort to break up the carbide-api mega-crate into
focused, independently buildable crates. The site-explorer, currently
living inside carbide-api, is the next planned extraction and depends
on a clean Redfish boundary — so the Redfish layer is split out first.

Moved out of crates/api/src/:

  redfish.rs     -> crates/redfish/src/libredfish/
  nv_redfish.rs  -> crates/redfish/src/nv_redfish/

The new carbide-redfish crate keeps both backends as separate
top-level modules, which also supports the in-flight libredfish ->
nv-redfish migration.

Public surface:

- Concrete RedfishClientPoolImpl is no longer exported; callers
  construct a pool via carbide_redfish::libredfish::new_pool(),
  returning Arc<dyn RedfishClientPool>.
- RedfishAuth, RedfishClientCreationError, redact_password, and the
  RedfishSim test helpers are re-exported from the new crate.
- Test helpers are gated behind a `test-support` cargo feature.

carbide-api now depends on carbide-redfish instead of owning the
Redfish code. All internal call sites (api, site_explorer, state
controllers, tests) and bmc-explorer-cli are updated to import from
carbide_redfish. No behavioral changes.
