+++ crates/api/src/preingestion_manager/mod.rs
@@ -21,6 +21,7 @@ use std::io;
 use std::sync::Arc;
 use std::time::Duration;
 
+use carbide_redfish::libredfish::{RedfishClientCreationError, RedfishClientPool};
 use chrono::{DateTime, Utc};
 use db::work_lock_manager::WorkLockManagerHandle;
 use db::{DatabaseError, WithTransaction};
@@ -45,7 +46,6 @@ use utils::periodic_timer::PeriodicTimer;
 use crate::cfg::file::{CarbideConfig, FirmwareConfig, FirmwareGlobal};
 use crate::firmware_downloader::FirmwareDownloader;
 use crate::preingestion_manager::metrics::PreingestionMetrics;
-use crate::redfish::{RedfishClientCreationError, RedfishClientPool};
 use crate::site_explorer::EndpointExplorer;
 use crate::{CarbideError, CarbideResult};
