+++ crates/api/src/site_explorer/bmc_endpoint_explorer.rs
@@ -21,6 +21,8 @@ use std::sync::Arc;
 use std::sync::atomic::{AtomicBool, Ordering};
 
 use carbide_ipmi::IPMITool;
+use carbide_redfish::libredfish::RedfishClientPool;
+use carbide_redfish::nv_redfish::NvRedfishClientPool;
 use forge_secrets::credentials::{CredentialManager, Credentials};
 use libredfish::model::oem::nvidia_dpu::NicMode;
 use libredfish::model::service_root::RedfishVendor;
@@ -38,8 +40,6 @@ use super::credentials::{CredentialClient, get_bmc_root_credential_key};
 use super::metrics::SiteExplorationMetrics;
 use super::redfish::RedfishClient;
 use crate::cfg::file::SiteExplorerExploreMode;
-use crate::nv_redfish::NvRedfishClientPool;
-use crate::redfish::RedfishClientPool;
 use crate::site_explorer::EndpointExplorer;
 
 const UNIFIED_PREINGESTION_BFB_PATH: &str =
