+++ crates/api/src/state_controller/machine/io.rs
@@ -33,7 +33,7 @@ use crate::state_controller::io::StateControllerIO;
 use crate::state_controller::machine::context::MachineStateHandlerContextObjects;
 use crate::state_controller::machine::metrics::MachineMetricsEmitter;
 
-// This should be updated on each new model introdunction
+// This should be updated on each new model introduction
 pub const CURRENT_STATE_MODEL_VERSION: i16 = 2;
 
 /// State Controller IO implementation for Machines
