+++ crates/bmc-explorer-cli/Cargo.toml
@@ -11,6 +11,7 @@ authors.workspace = true
 carbide-api = { path = "../api", default-features = false }
 carbide-secrets = { path = "../secrets" }
 carbide-ipmi = { path = "../ipmi" }
+carbide-redfish = { path = "../redfish" }
 # DO NOT PUT DEPENDENCIES OTHER THAN LOCAL DEPS HERE, THEY SHOULD ALL HAVE 'path =' IN THEM
 
 #these are alphabetized
