+++ crates/bmc-explorer-cli/src/main.rs
@@ -18,9 +18,8 @@ use std::sync::Arc;
 use std::time::Instant;
 
 use arc_swap::ArcSwap;
-use carbide::{
-    BmcEndpointExplorer, NvRedfishClientPool, RedfishClientPoolImpl, SiteExplorerExploreMode,
-};
+use carbide::{BmcEndpointExplorer, SiteExplorerExploreMode};
+use carbide_redfish::nv_redfish::NvRedfishClientPool;
 use clap::Parser;
 use forge_secrets::credentials::{Credentials, TestCredentialManager};
 use mac_address::MacAddress;
@@ -80,11 +79,11 @@ async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let proxy_address = Arc::new(ArcSwap::new(None.into()));
     let credential_provider = Arc::new(TestCredentialManager::new(fallback_credentials.clone()));
 
-    let redfish_client_pool = Arc::new(RedfishClientPoolImpl::new(
+    let redfish_client_pool = carbide_redfish::libredfish::new_pool(
         credential_provider.clone(),
         rf_pool,
         proxy_address.clone(),
-    ));
+    );
     let mode = match args.mode.as_str() {
         "libredfish" => SiteExplorerExploreMode::LibRedfish,
         "nv-redfish" => SiteExplorerExploreMode::NvRedfish,
